COM-2842: export bills and credit notes
