+++ src/helpers/historyExport.js
@@ -1,9 +1,8 @@
 const get = require('lodash/get');
 const pick = require('lodash/pick');
-const uniqBy = require('lodash/uniqBy');
-const groupBy = require('lodash/groupBy');
 const { keyBy } = require('lodash');
 const {
+  NO_DATA,
   NEVER,
   EVENT_TYPE_LIST,
   REPETITION_FREQUENCY_TYPE_LIST,
@@ -23,15 +22,6 @@ const {
   TIMESTAMPING_ACTION_TYPE_LIST,
   MANUAL_TIME_STAMPING_REASONS,
   EVENT_TRANSPORT_MODE_LIST,
-  INTRA,
-  ON_SITE,
-  REMOTE,
-  STEP_TYPES,
-  EXPECTATIONS,
-  END_OF_COURSE,
-  OPEN_QUESTION,
-  SURVEY,
-  QUESTION_ANSWER,
   TIME_STAMPING_ACTIONS,
 } = require('./constants');
 const DatesHelper = require('./dates');
@@ -40,26 +30,16 @@ const { CompaniDuration } = require('./dates/companiDurations');
 const UtilsHelper = require('./utils');
 const NumbersHelper = require('./numbers');
 const DraftPayHelper = require('./draftPay');
-const CourseBillHelper = require('./courseBills');
-const CourseHelper = require('./courses');
-const AttendanceSheet = require('../models/AttendanceSheet');
 const DistanceMatrixHelper = require('./distanceMatrix');
 const Event = require('../models/Event');
 const Bill = require('../models/Bill');
 const CreditNote = require('../models/CreditNote');
 const Contract = require('../models/Contract');
-const CourseSmsHistory = require('../models/CourseSmsHistory');
-const CourseSlot = require('../models/CourseSlot');
 const Pay = require('../models/Pay');
 const Payment = require('../models/Payment');
 const FinalPay = require('../models/FinalPay');
-const CourseRepository = require('../repositories/CourseRepository');
 const EventRepository = require('../repositories/EventRepository');
 const UserRepository = require('../repositories/UserRepository');
-const QuestionnaireHistory = require('../models/QuestionnaireHistory');
-const Questionnaire = require('../models/Questionnaire');
-
-const NO_DATA = 'Aucune donnée sur la periode selectionnée';
 
 const workingEventExportHeader = [
   'Type',
@@ -672,206 +652,6 @@ exports.exportPaymentsHistory = async (startDate, endDate, credentials) => {
   return rows;
 };
 
-const getEndOfCourse = (slotsGroupedByDate, slotsToPlan) => {
-  if (slotsToPlan.length) return 'à planifier';
-  if (slotsGroupedByDate.length) {
-    const lastDate = slotsGroupedByDate.length - 1;
-    const lastSlot = slotsGroupedByDate[lastDate].length - 1;
-    return CompaniDate(slotsGroupedByDate[lastDate][lastSlot].endDate).format('dd/LL/yyyy HH:mm:ss');
-  }
-  return '';
-};
-
-const getStartOfCourse = slotsGroupedByDate => (slotsGroupedByDate.length
-  ? CompaniDate(slotsGroupedByDate[0][0].startDate).format('dd/LL/yyyy HH:mm:ss')
-  : '');
-
-const isSlotInInterval = (slot, startDate, endDate) => CompaniDate(slot.startDate).isAfter(startDate) &&
-  CompaniDate(slot.endDate).isBefore(endDate);
-
-const getAttendancesCountInfos = (course) => {
-  const attendances = course.slots.map(slot => slot.attendances).flat();
-  const courseTraineeList = course.trainees.map(trainee => trainee._id);
-  const subscribedAttendances = attendances
-    .filter(attendance => UtilsHelper.doesArrayIncludeId(courseTraineeList, attendance.trainee))
-    .length;
-
-  const upComingSlots = course.slots.filter(slot => CompaniDate().isBefore(slot.startDate)).length;
-  const attendancesToCome = upComingSlots * course.trainees.length;
-
-  const unsubscribedTrainees = uniqBy(attendances.map(a => a.trainee), trainee => trainee.toString())
-    .filter(attendanceTrainee => !UtilsHelper.doesArrayIncludeId(courseTraineeList, attendanceTrainee))
-    .length;
-
-  return {
-    subscribedAttendances,
-    unsubscribedAttendances: attendances.length - subscribedAttendances,
-    absences: (course.slots.length * course.trainees.length) - subscribedAttendances - attendancesToCome,
-    unsubscribedTrainees,
-    pastSlots: course.slots.length - upComingSlots,
-  };
-};
-
-const getBillsInfos = (bills) => {
-  const courseBillsWithoutCreditNote = bills.filter(bill => !bill.courseCreditNote);
-  const payer = courseBillsWithoutCreditNote.map(bill => get(bill, 'payer.name')).toString();
-  const isBilled = courseBillsWithoutCreditNote.map(bill => (bill.billedAt ? 'Oui' : 'Non')).toString();
-
-  const validatedBill = courseBillsWithoutCreditNote.find(bill => bill.billedAt);
-  const { netInclTaxes, paid, total } = validatedBill
-    ? CourseBillHelper.computeAmounts(validatedBill)
-    : { netInclTaxes: '', paid: '', total: '' };
-
-  return { isBilled, payer, netInclTaxes, paid, total };
-};
-
-exports.exportCourseHistory = async (startDate, endDate, credentials) => {
-  const courses = await CourseRepository.findCoursesForExport(startDate, endDate, credentials);
-
-  const filteredCourses = courses
-    .filter(course => !course.slots.length || course.slots.some(slot => isSlotInInterval(slot, startDate, endDate)));
-
-  const courseIds = filteredCourses.map(course => course._id);
-  const [questionnaireHistories, smsList, attendanceSheetList] = await Promise.all([
-    QuestionnaireHistory
-      .find({ course: { $in: courseIds }, select: 'course questionnaire' })
-      .populate({ path: 'questionnaire', select: 'type' })
-      .lean(),
-    CourseSmsHistory.find({ course: { $in: courseIds }, select: 'course' }).lean(),
-    AttendanceSheet.find({ course: { $in: courseIds }, select: 'course' }).lean(),
-  ]);
-
-  const rows = [];
-  const groupedSms = groupBy(smsList, 'course');
-  const grouppedAttendanceSheets = groupBy(attendanceSheetList, 'course');
-  const groupedCourseQuestionnaireHistories = groupBy(questionnaireHistories, 'course');
-
-  for (const course of filteredCourses) {
-    const slotsGroupedByDate = CourseHelper.groupSlotsByDate(course.slots);
-    const smsCount = (groupedSms[course._id] || []).length;
-    const attendanceSheets = (grouppedAttendanceSheets[course._id] || []).length;
-    const {
-      subscribedAttendances,
-      unsubscribedAttendances,
-      absences,
-      unsubscribedTrainees,
-      pastSlots,
-    } = getAttendancesCountInfos(course);
-
-    const courseQuestionnaireHistories = groupedCourseQuestionnaireHistories[course._id] || [];
-    const expectactionQuestionnaireAnswers = courseQuestionnaireHistories
-      .filter(qh => qh.questionnaire.type === EXPECTATIONS)
-      .length;
-    const endQuestionnaireAnswers = courseQuestionnaireHistories
-      .filter(qh => qh.questionnaire.type === END_OF_COURSE)
-      .length;
-
-    const traineeProgressList = course.trainees
-      .map(trainee => CourseHelper.getTraineeElearningProgress(trainee._id, course.subProgram.steps))
-      .filter(trainee => trainee.progress.eLearning >= 0)
-      .map(trainee => trainee.progress.eLearning);
-    const combinedElearningProgress = traineeProgressList.reduce((acc, value) => acc + value, 0);
-
-    const { isBilled, payer, netInclTaxes, paid, total } = getBillsInfos(course.bills);
-
-    rows.push({
-      Identifiant: course._id,
-      Type: course.type,
-      Payeur: payer || '',
-      Structure: course.type === INTRA ? get(course, 'company.name') : '',
-      Programme: get(course, 'subProgram.program.name') || '',
-      'Sous-Programme': get(course, 'subProgram.name') || '',
-      'Infos complémentaires': course.misc,
-      Formateur: UtilsHelper.formatIdentity(get(course, 'trainer.identity') || '', 'FL'),
-      'Référent Compani': UtilsHelper.formatIdentity(get(course, 'salesRepresentative.identity') || '', 'FL'),
-      'Contact pour la formation': UtilsHelper.formatIdentity(get(course, 'contact.identity') || '', 'FL'),
-      'Nombre d\'inscrits': get(course, 'trainees.length'),
-      'Nombre de dates': slotsGroupedByDate.length,
-      'Nombre de créneaux': get(course, 'slots.length'),
-      'Nombre de créneaux à planifier': get(course, 'slotsToPlan.length'),
-      'Durée Totale': UtilsHelper.getTotalDurationForExport(course.slots),
-      'Nombre de SMS envoyés': smsCount,
-      'Nombre de personnes connectées à l\'app': course.trainees
-        .filter(trainee => trainee.firstMobileConnection).length,
-      'Complétion eLearning moyenne': traineeProgressList.length
-        ? UtilsHelper.formatFloatForExport(combinedElearningProgress / course.trainees.length)
-        : '',
-      'Nombre de réponses au questionnaire de recueil des attentes': expectactionQuestionnaireAnswers,
-      'Nombre de réponses au questionnaire de satisfaction': endQuestionnaireAnswers,
-      'Date de démarrage souhaitée': course.estimatedStartDate
-        ? CompaniDate(course.estimatedStartDate).format('dd/LL/yyyy')
-        : '',
-      'Début de formation': getStartOfCourse(slotsGroupedByDate),
-      'Fin de formation': getEndOfCourse(slotsGroupedByDate, course.slotsToPlan),
-      'Nombre de feuilles d\'émargement chargées': attendanceSheets,
-      'Nombre de présences': subscribedAttendances,
-      'Nombre d\'absences': absences,
-      'Nombre de stagiaires non prévus': unsubscribedTrainees,
-      'Nombre de présences non prévues': unsubscribedAttendances,
-      Avancement: UtilsHelper.formatFloatForExport(pastSlots / (course.slots.length + course.slotsToPlan.length)),
-      Facturée: isBilled,
-      'Montant facturé': netInclTaxes,
-      'Montant réglé': paid,
-      Solde: total,
-    });
-  }
-
-  return rows.length ? [Object.keys(rows[0]), ...rows.map(d => Object.values(d))] : [[NO_DATA]];
-};
-
-const getAddress = (slot) => {
-  if (get(slot, 'step.type') === ON_SITE) return get(slot, 'address.fullAddress') || '';
-  if (get(slot, 'step.type') === REMOTE) return slot.meetingLink || '';
-
-  return '';
-};
-
-exports.exportCourseSlotHistory = async (startDate, endDate) => {
-  const courseSlots = await CourseSlot.find({ startDate: { $lte: endDate }, endDate: { $gte: startDate } })
-    .populate({ path: 'step', select: 'type name' })
-    .populate({
-      path: 'course',
-      select: 'type trainees misc subProgram company',
-      populate: [
-        { path: 'company', select: 'name' },
-        { path: 'subProgram', select: 'program', populate: [{ path: 'program', select: 'name' }] },
-      ],
-    })
-    .populate({ path: 'attendances' })
-    .lean();
-
-  const rows = [];
-
-  for (const slot of courseSlots) {
-    const slotDuration = UtilsHelper.getDurationForExport(slot.startDate, slot.endDate);
-    const subscribedAttendances = slot.attendances
-      .filter(attendance => UtilsHelper.doesArrayIncludeId(slot.course.trainees, attendance.trainee))
-      .length;
-
-    const courseName = get(slot, 'course.type') === INTRA
-      ? `${slot.course.company.name} - ${slot.course.subProgram.program.name} - ${slot.course.misc}`
-      : `${slot.course.subProgram.program.name} - ${slot.course.misc}`;
-
-    rows.push({
-      'Id Créneau': slot._id,
-      'Id Formation': slot.course._id,
-      Formation: courseName,
-      Étape: get(slot, 'step.name') || '',
-      Type: STEP_TYPES[get(slot, 'step.type')] || '',
-      'Date de création': CompaniDate(slot.createdAt).format('dd/LL/yyyy HH:mm:ss') || '',
-      'Date de début': CompaniDate(slot.startDate).format('dd/LL/yyyy HH:mm:ss') || '',
-      'Date de fin': CompaniDate(slot.endDate).format('dd/LL/yyyy HH:mm:ss') || '',
-      Durée: slotDuration,
-      Adresse: getAddress(slot),
-      'Nombre de présences': subscribedAttendances,
-      'Nombre d\'absences': slot.course.trainees.length - subscribedAttendances,
-      'Nombre de présences non prévues': slot.attendances.length - subscribedAttendances,
-    });
-  }
-
-  return rows.length ? [Object.keys(rows[0]), ...rows.map(d => Object.values(d))] : [[NO_DATA]];
-};
-
 exports.exportTransportsHistory = async (startDate, endDate, credentials) => {
   const rows = [];
   const events = await EventRepository.getEventsByDayAndAuxiliary(
@@ -927,78 +707,6 @@ exports.exportTransportsHistory = async (startDate, endDate, credentials) => {
         });
       }
     }
-  }
-
-  return rows.length ? [Object.keys(rows[0]), ...rows.map(d => Object.values(d))] : [[NO_DATA]];
-};
-
-const _findAnswerText = (answers, answerId) => {
-  const answer = answers.find(qa => UtilsHelper.areObjectIdsEquals(qa._id, answerId));
-
-  return answer ? answer.text : '';
-};
-
-const _getAnswerForExport = (questionnaireCard, questionnaireHistoryAnswersList) => {
-  const qAnswer = questionnaireHistoryAnswersList
-    .find(qa => UtilsHelper.areObjectIdsEquals(qa.card._id, questionnaireCard._id));
-
-  return qAnswer
-    ? qAnswer.answerList
-      .map(a => (UtilsHelper.isStringedObjectId(a) ? _findAnswerText(qAnswer.card.qcAnswers, a) : a))
-      .join()
-    : '';
-};
-
-exports.exportEndOfCourseQuestionnaireHistory = async (startDate, endDate) => {
-  const rows = [];
-
-  const endOfCourseQuestionnaire = await Questionnaire
-    .findOne({ type: END_OF_COURSE })
-    .populate({ path: 'cards', select: 'question template' })
-    .populate({
-      path: 'histories',
-      match: { createdAt: { $gte: startDate, $lte: endDate } },
-      populate: [
-        {
-          path: 'course',
-          select: 'subProgram',
-          populate: [
-            { path: 'subProgram', select: 'name program', populate: { path: 'program', select: 'name' } },
-            { path: 'trainer', select: 'identity' },
-          ],
-        },
-        {
-          path: 'user',
-          select: 'identity local.email contact.phone company',
-          populate: { path: 'company', populate: { path: 'company', select: 'name' } },
-        },
-        { path: 'questionnaireAnswersList.card', select: 'qcAnswers' },
-      ],
-    })
-    .lean({ virtuals: true });
-
-  for (const qHistory of endOfCourseQuestionnaire.histories) {
-    const questionsAnswers = endOfCourseQuestionnaire.cards
-      .filter(card => [OPEN_QUESTION, SURVEY, QUESTION_ANSWER].includes(card.template))
-      .reduce((acc, card) => ({
-        ...acc,
-        [card.question]: _getAnswerForExport(card, qHistory.questionnaireAnswersList),
-      }), {});
-
-    const row = {
-      'Id formation': qHistory.course._id,
-      Programme: get(qHistory, 'course.subProgram.program.name') || '',
-      'Sous-programme': get(qHistory, 'course.subProgram.name'),
-      'Prénom Nom intervenant(e)': UtilsHelper.formatIdentity(get(qHistory, 'course.trainer.identity') || '', 'FL'),
-      Structure: get(qHistory, 'user.company.name'),
-      'Date de réponse': CompaniDate(qHistory.createdAt).format('dd/LL/yyyy HH:mm:ss'),
-      'Prénom Nom répondant(e)': UtilsHelper.formatIdentity(get(qHistory, 'user.identity') || '', 'FL'),
-      'Mail répondant(e)': get(qHistory, 'user.local.email'),
-      'Numéro de tél répondant(e)': get(qHistory, 'user.contact.phone') || '',
-      ...questionsAnswers,
-    };
-
-    rows.push(row);
   }
 
   return rows.length ? [Object.keys(rows[0]), ...rows.map(d => Object.values(d))] : [[NO_DATA]];
